+++ Kinopoisk/Source/Constants/CommonMetrics.swift
@@ -10,4 +10,7 @@ import UIKit
 enum CommonMetrics {
     static let leadingOffset: CGFloat = 20
     static let trailingOffset: CGFloat = -20
+    
+    static let defaultFontSize: CGFloat = 14
+    static let defaultNumberOfLines: Int = 1
 }
