+++ Kinopoisk/Source/Extensions/UIStackView+Extension.swift
@@ -9,7 +9,7 @@ import UIKit
 
 // MARK: - StackView creation
 
-public extension UIStackView {
+extension UIStackView {
     static func createStackView(axis: NSLayoutConstraint.Axis,
                                 distribution: UIStackView.Distribution,
                                 alignment: UIStackView.Alignment,
