+++ Kinopoisk/Source/Scenes/FilmsList/View/FilmsListTableViewCell.swift
@@ -38,10 +38,10 @@ class FilmsListTableViewCell: UITableViewCell {
     private lazy var textStackView = UIStackView.createStackView(axis: .vertical, distribution: .equalSpacing, alignment: .leading)
     
     // MARK: Labels
-    private lazy var titleLabel = createLabel(fontWeight: .bold, fontSize: Metric.titleFontSize, numberOfLines: Metric.titleNumberOfLines)
-    private lazy var alternativeTitleLabel = createLabel(fontWeight: .regular, textColor: .secondaryLabel)
-    private lazy var yearLabel = createLabel(fontWeight: .regular, textColor: .secondaryLabel)
-    private lazy var ratingLabel = createLabel(fontWeight: .bold, textColor: .white)
+    private lazy var titleLabel = UILabel.createLabel(fontWeight: .bold, fontSize: Metric.titleFontSize, numberOfLines: Metric.titleNumberOfLines)
+    private lazy var alternativeTitleLabel = UILabel.createLabel(fontWeight: .regular, textColor: .secondaryLabel)
+    private lazy var yearLabel = UILabel.createLabel(fontWeight: .regular, textColor: .secondaryLabel)
+    private lazy var ratingLabel = UILabel.createLabel(fontWeight: .bold, textColor: .white)
     
     // MARK: Auxiliary views
     private lazy var ratingView: UIView = {
@@ -110,20 +110,6 @@ class FilmsListTableViewCell: UITableViewCell {
             make.center.equalToSuperview()
         }
     }
-    
-    // MARK: - Private functions
-    
-    private func createLabel(fontWeight: UIFont.Weight,
-                             fontSize: CGFloat = Metric.defaultFontSize,
-                             numberOfLines: Int = Metric.defaultNumberOfLines,
-                             textColor: UIColor = .label) -> UILabel {
-        let label = UILabel()
-        label.font = .systemFont(ofSize: fontSize, weight: fontWeight)
-        label.numberOfLines = numberOfLines
-        label.textColor = textColor
-        
-        return label
-    }
 }
 
 // MARK: - Constants
@@ -139,10 +125,7 @@ extension FilmsListTableViewCell {
         static let ratingWidth: CGFloat = 32
         static let ratingCornerRadius: CGFloat = 3
         
-        static let defaultFontSize: CGFloat = 14
         static let titleFontSize: CGFloat = 16
-        
-        static let defaultNumberOfLines: Int = 1
         static let titleNumberOfLines: Int = 2
     }
 }
